requirements.txt
Django==2.2
django-elasticsearch-dsl==0.5.1
django-elasticsearch-dsl-drf==0.17.6
django-grappelli==2.12.2
django-guardian==1.5.0
django-haystack==2.8.1
django-jalali==3.1.0
django-jalali-date==0.2.9
django-nine==0.2.2
django-rest-auth==0.9.5
django-rest-guardian==1.1.1
django-rest-registration==0.4.3
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
drf-haystack==1.8.5
elasticsearch==7.0.1
elasticsearch-dsl==7.0.0
gunicorn==19.9.0
jdatetime==3.2.0
Pillow==6.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
simplejwt==2.0.1
six==1.12.0
sqlparse==0.3.0
typing==3.6.6
urllib3==1.25.3
utils==0.9.0
